v1/requirements.txt
geopandas==1.0.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
matplotlib-venn==1.1.1
nltk==3.9.1
numpy==1.26.4
pandas==2.2.2
pandas-datareader==0.10.0
pandas-gbq==0.26.1
pandas-stubs==2.2.2.240909
scikit-learn==1.6.1
seaborn==0.13.2
sklearn-pandas==2.2.0
tensorflow==2.18.0
tensorflow-datasets==4.9.7
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-metadata==1.16.1
tensorflow-probability==0.25.0
tensorflow-text==2.18.1
